Update README.md
STARTERS 119 DIVISION 4 (RATED) - DETAILS

CodeChef: A Platform for Aspiring Programmers
CodeChef was created as a platform to help programmers make it big in the world of algorithms, computer programming and programming contests. At CodeChef, we work hard to revive the geek in you by hosting a programming contest at the start of the month and another smaller programming challenge in the middle of the month.